added derived_from_files option
Added support for explicitly decleared derived_from files. The user can provide a comma separated list of files that will be added to the 'derived_from' list. These files can be any (no check is done by the script). 
This is designed to support adding s3 objects directly in the 'derived_from' list